refactor: 패키지 구조 변경 및 fake class 네이밍 수정
