+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/fixture/FakeManualHandlerAdapter.java
@@ -1,4 +1,4 @@
-package samples;
+package com.interface21.webmvc.servlet.mvc.tobe.fixture;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -6,7 +6,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
 
-public class TestManualHandlerAdapter implements HandlerAdapter {
+public class FakeManualHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(Object handler) {
